Move SpaceAfterFinal to sniperPrinter
